requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
csrf==0.1b1
Django==4.1.4
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.13
